# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TTDMI - Tourism Trust & Digital Maturity Index API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /operators:
    get:
      operationId: listOperators
      tags: [operators]
      summary: List tourism operators with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: operator_type
          in: query
          schema:
            type: string
        - name: niche
          in: query
          schema:
            type: string
        - name: min_score
          in: query
          schema:
            type: number
        - name: max_score
          in: query
          schema:
            type: number
        - name: verified
          in: query
          schema:
            type: boolean
        - name: level
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sort_by
          in: query
          schema:
            type: string
            default: ttdmi_score
      responses:
        "200":
          description: List of operators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorListResponse"

    post:
      operationId: createOperator
      tags: [operators]
      summary: Create a new tourism operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /operators/{id}:
    get:
      operationId: getOperator
      tags: [operators]
      summary: Get operator details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Operator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorDetail"
        "404":
          description: Not found

    patch:
      operationId: updateOperator
      tags: [operators]
      summary: Update operator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"

    delete:
      operationId: deleteOperator
      tags: [operators]
      summary: Delete operator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /operators/{id}/score:
    get:
      operationId: getOperatorScore
      tags: [operators]
      summary: Get full TTDMI score breakdown for an operator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TtdmiScore"

    post:
      operationId: recalculateScore
      tags: [operators]
      summary: Recalculate TTDMI score for an operator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recalculated score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TtdmiScore"
...
components:
  schemas:
...
    Operator:
      type: object
      required: [id, legal_name, commercial_name, region, operator_type, ttdmi_score, level, verified, created_at]
      properties:
        id:
          type: integer
        legal_name:
          type: string
        commercial_name:
          type: string
        ruc:
          type: ["string", "null"]
        region:
          type: string
        province:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        operator_type:
          type: string
        niche:
          type: ["string", "null"]
        languages:
          type: array
          items:
            type: string
        verified:
          type: boolean
        ttdmi_score:
          type: number
        level:
          type: string
        level_label:
          type: string
        website:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        rank_nacional:
          type: ["integer", "null"]
        created_at:
          type: string
    OperatorDetail:
      type: object
      required: [id, legal_name, commercial_name, region, operator_type, ttdmi_score, level, verified, created_at, score_breakdown]
      properties:
        id:
          type: integer
        legal_name:
          type: string
        commercial_name:
          type: string
        ruc:
          type: ["string", "null"]
        region:
          type: string
        province:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        operator_type:
          type: string
        niche:
          type: ["string", "null"]
        languages:
          type: array
          items:
            type: string
        verified:
          type: boolean
        ttdmi_score:
          type: number
        level:
          type: string
        level_label:
          type: string
        website:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        rank_nacional:
          type: ["integer", "null"]
        created_at:
          type: string
        score_breakdown:
          $ref: "#/components/schemas/TtdmiScore"
        social_profiles:
          type: array
          items:
            $ref: "#/components/schemas/SocialProfile"
        reviews_summary:
          $ref: "#/components/schemas/ReviewsSummary"
        booking_capabilities:
          $ref: "#/components/schemas/BookingCapabilities"
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
    OperatorListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Operator"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OperatorInput:
      type: object
      required: [legal_name, commercial_name, region, operator_type]
      properties:
        legal_name:
          type: string
        commercial_name:
          type: string
        ruc:
          type: string
        region:
          type: string
        province:
          type: string
        district:
          type: string
        operator_type:
          type: string
        niche:
          type: string
        languages:
          type: array
          items:
            type: string
        website:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
    OperatorUpdate:
      type: object
      properties:
        legal_name:
          type: string
        commercial_name:
          type: string
        ruc:
          type: string
        region:
          type: string
        province:
          type: string
        district:
          type: string
        operator_type:
          type: string
        niche:
          type: string
        languages:
          type: array
          items:
            type: string
        website:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        verified:
          type: boolean
    TtdmiScore:
      type: object
      required: [operator_id, total_score, reviews_score, social_score, website_score, seo_score, engagement_score, formalidad_score, conversion_score, freshness_score, network_score, validation_factor, level, updated_at]
      properties:
        operator_id:
          type: integer
        total_score:
          type: number
        reviews_score:
          type: number
        social_score:
          type: number
        website_score:
          type: number
        seo_score:
          type: number
        engagement_score:
          type: number
        formalidad_score:
          type: number
        conversion_score:
          type: number
        freshness_score:
          type: number
        network_score:
          type: number
        validation_factor:
          type: number
        level:
          type: string
        level_label:
          type: string
        score_details:
          type: object
          additionalProperties: true
        updated_at:
          type: string
    SocialProfile:
      type: object
      required: [id, platform, followers, engagement_rate, verified]
      properties:
        id:
          type: integer
        platform:
          type: string
        handle:
          type: ["string", "null"]
        followers:
          type: integer
        engagement_rate:
          type: number
        verified:
          type: boolean
        last_post_date:
          type: ["string", "null"]
        posts_per_week:
          type: ["number", "null"]
    ReviewsSummary:
      type: object
      required: [average_rating, total_reviews, authenticity_score]
      properties:
        average_rating:
          type: number
        total_reviews:
          type: integer
        authenticity_score:
          type: number
        sources:
          type: array
          items:
            $ref: "#/components/schemas/ReviewSource"
        sentiment_score:
          type: number
        recency_score:
          type: number
    ReviewSource:
      type: object
      required: [source, count, avg_rating]
      properties:
        source:
          type: string
        count:
          type: integer
        avg_rating:
          type: number
    BookingCapabilities:
      type: object
      required: [booking_engine, online_payment, whatsapp_cta, chatbot, lead_form]
      properties:
        booking_engine:
          type: boolean
        online_payment:
          type: boolean
        whatsapp_cta:
          type: boolean
        chatbot:
          type: boolean
        lead_form:
          type: boolean
        response_time_hours:
          type: ["number", "null"]
        ota_presence:
          type: array
          items:
            type: string
    Certification:
      type: object
      required: [id, name, issuer, verified]
      properties:
        id:
          type: integer
        name:
          type: string
        issuer:
          type: string
        verified:
          type: boolean
        expiry_date:
          type: ["string", "null"]
...