# lib/api-zod/src/generated/types/bookingCapabilities.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTDMI - Tourism Trust & Digital Maturity Index API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingCapabilities {
  booking_engine: boolean;
  online_payment: boolean;
  whatsapp_cta: boolean;
  chatbot: boolean;
  lead_form: boolean;
  /** @nullable */
  response_time_hours?: number | null;
  ota_presence?: string[];
}