# lib/api-zod/src/generated/types/certification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTDMI - Tourism Trust & Digital Maturity Index API
 * OpenAPI spec version: 0.1.0
 */

export interface Certification {
  id: number;
  name: string;
  issuer: string;
  verified: boolean;
  /** @nullable */
  expiry_date?: string | null;
}