# lib/api-zod/src/generated/types/operator.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTDMI - Tourism Trust & Digital Maturity Index API
 * OpenAPI spec version: 0.1.0
 */

export interface Operator {
  id: number;
  legal_name: string;
  commercial_name: string;
  /** @nullable */
  ruc?: string | null;
  region: string;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  district?: string | null;
  operator_type: string;
  /** @nullable */
  niche?: string | null;
  languages?: string[];
  verified: boolean;
  ttdmi_score: number;
  level: string;
  level_label?: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  rank_nacional?: number | null;
  created_at: string;
}