# lib/api-zod/src/generated/types/operatorDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTDMI - Tourism Trust & Digital Maturity Index API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingCapabilities } from './bookingCapabilities';
import type { Certification } from './certification';
import type { ReviewsSummary } from './reviewsSummary';
import type { SocialProfile } from './socialProfile';
import type { TtdmiScore } from './ttdmiScore';

export interface OperatorDetail {
  id: number;
  legal_name: string;
  commercial_name: string;
  /** @nullable */
  ruc?: string | null;
  region: string;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  district?: string | null;
  operator_type: string;
  /** @nullable */
  niche?: string | null;
  languages?: string[];
  verified: boolean;
  ttdmi_score: number;
  level: string;
  level_label?: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  rank_nacional?: number | null;
  created_at: string;
  score_breakdown: TtdmiScore;
  social_profiles?: SocialProfile[];
  reviews_summary?: ReviewsSummary;
  booking_capabilities?: BookingCapabilities;
  certifications?: Certification[];
}