# lib/api-zod/src/generated/types/operatorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTDMI - Tourism Trust & Digital Maturity Index API
 * OpenAPI spec version: 0.1.0
 */

export interface OperatorInput {
  legal_name: string;
  commercial_name: string;
  ruc?: string;
  region: string;
  province?: string;
  district?: string;
  operator_type: string;
  niche?: string;
  languages?: string[];
  website?: string;
  description?: string;
  phone?: string;
  email?: string;
}